+++ src/Storage/Storage/Common/AzureStorageFile.cs
@@ -72,8 +72,33 @@ namespace Microsoft.WindowsAzure.Commands.Common.Storage.ResourceModel
         /// </summary>
         public global::Azure.Storage.Files.Shares.Models.ShareFileItem ListFileProperties { get; private set; }
 
+        /// <summary>
+        /// XSCL Track2 File Hardlink create returned properties
+        /// </summary>
+        public global::Azure.Storage.Files.Shares.Models.ShareFileInfo ShareFileInfo { get; private set; }
+
 
         private ShareClientOptions shareClientOptions { get; set; }
+
+        /// <summary>
+        /// Azure storage file constructor from Track2 list file item
+        /// </summary>
+        /// <param name="shareFileClient"></param>
+        /// <param name="storageContext"></param>
+        /// <param name="info"></param>
+        /// <param name="clientOptions"></param>
+        public AzureStorageFile(ShareFileClient shareFileClient, AzureStorageContext storageContext, ShareFileInfo info, ShareClientOptions clientOptions = null)
+        {
+            Name = shareFileClient.Name;
+            this.privateFileClient = shareFileClient;
+            if (info != null)
+            {
+                ShareFileInfo = info;
+                LastModified = info.LastModified;
+            }
+            Context = storageContext;
+            shareClientOptions = clientOptions;
+        }
 
         /// <summary>
         /// Azure storage file constructor from Track2 list file item